Refactored the MenuSystem slightly more.
There is kinda a "bug" that we'll have a hard time fixing...

... and that's the fact that the settings menu seems to need to be
exclusive.  It shouldn't have any other menus in the background
as doing so will cause some of the buttons to navigate away from
the settings menu and this can cause issues with using the menus
entirely.

I will probably alter the Settings menu in the future to be exclusive..
well.. specifically the Controls settings as it's the part we can't
control.  (ReWired generates the buttons automatically at runtime).