Fixed some slight navigation flow issues.
NOTE:  I'm a little saddened by our MenuManager..  It's not actually
doing what it was originally designed to handle and is instead simply
acting as a Menu container from which we can declare "display this".

Oh well..  Too much time has already been put into it.  Best to just use
it as is since it works.