InitializePLEMenu Integration
Added new prefab to the Menu Manager (the first of many to come)
Changed the "Camera" in the PlayerLevelEditor namespace to be called "PLECameraController" to get rid of ambiguity with UnityEngine.Camera.
Added a null check in the toolkit for drawing those preview blocks, may not be needed as this stuff should be moved out to the calling function. 
Weird draw bug when drawing the preview block positions.